requirements.txt
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.10
idna==3.3
iniconfig==1.1.1
isodate==0.6.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pydantic==1.9.0
pyparsing==3.0.6
pytest==6.2.5
python-dotenv==0.19.2
PyYAML==6.0
rdflib==6.1.1
rdfogm==0.0.1
requests==2.27.1
requests-toolbelt==0.9.1
six==1.16.0
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.8
